+++ src/werkzeug/test.py
@@ -1,6 +1,7 @@
 from __future__ import annotations
 
 import dataclasses
+import json
 import mimetypes
 import sys
 import typing as t
@@ -11,6 +12,7 @@ from itertools import chain
 from random import random
 from tempfile import TemporaryFile
 from time import time
+from types import TracebackType
 from urllib.parse import unquote
 from urllib.parse import urlsplit
 from urllib.parse import urlunsplit
@@ -172,9 +174,6 @@ def _iter_data(data: t.Mapping[str, t.Any]) -> t.Iterator[tuple[str, t.Any]]:
                 yield key, value
 
 
-_TAnyMultiDict = t.TypeVar("_TAnyMultiDict", bound="MultiDict[t.Any, t.Any]")
-
-
 class EnvironBuilder:
     """This class can be used to conveniently create a WSGI environment
     for testing purposes.  It can be used to quickly create WSGI environments
@@ -185,28 +184,6 @@ class EnvironBuilder:
     :meth:`Client.open`).  Because of this most of the functionality is
     available through the constructor alone.
 
-    Files and regular form data can be manipulated independently of each
-    other with the :attr:`form` and :attr:`files` attributes, but are
-    passed with the same argument to the constructor: `data`.
-
-    `data` can be any of these values:
-
-    -   a `str` or `bytes` object: The object is converted into an
-        :attr:`input_stream`, the :attr:`content_length` is set and you have to
-        provide a :attr:`content_type`.
-    -   a `dict` or :class:`MultiDict`: The keys have to be strings. The values
-        have to be either any of the following objects, or a list of any of the
-        following objects:
-
-        -   a :class:`file`-like object:  These are converted into
-            :class:`FileStorage` objects automatically.
-        -   a `tuple`:  The :meth:`~FileMultiDict.add_file` method is called
-            with the key and the unpacked `tuple` items as positional
-            arguments.
-        -   a `str`:  The string is set as form data for the associated key.
-    -   a file-like object: The object content is loaded in memory and then
-        handled like a regular `str` or a `bytes`.
-
     :param path: the path of the request.  In the WSGI environment this will
                  end up as `PATH_INFO`.  If the `query_string` is not defined
                  and there is a question mark in the `path` everything after
@@ -214,12 +191,10 @@ class EnvironBuilder:
     :param base_url: the base URL is a URL that is used to extract the WSGI
                      URL scheme, host (server name + server port) and the
                      script root (`SCRIPT_NAME`).
-    :param query_string: an optional string or dict with URL parameters.
+    :param query_string: A :class:`dict` or :class:`.MultiDict` to encode as the
+        query string of the URL, which sets :attr:`args`. Or a string, which
+        sets :attr:`query_string`, in which case :attr:`args` cannot be used.
     :param method: the HTTP method to use, defaults to `GET`.
-    :param input_stream: an optional input stream.  Do not specify this and
-                         `data`.  As soon as an input stream is set you can't
-                         modify :attr:`args` and :attr:`files` unless you
-                         set the :attr:`input_stream` to `None` again.
     :param content_type: The content type for the request.  As of 0.5 you
                          don't have to provide this when specifying files
                          and form data via `data`.
@@ -232,8 +207,12 @@ class EnvironBuilder:
     :param multiprocess: controls `wsgi.multiprocess`.  Defaults to `False`.
     :param run_once: controls `wsgi.run_once`.  Defaults to `False`.
     :param headers: an optional list or :class:`Headers` object of headers.
-    :param data: a string or dict of form data or a file-object.
-                 See explanation above.
+    :param data: A dict of form and file data to encode as the body of the
+        request; file values can be an IO object, ``(stream, filename)``,
+        ``(stream, filename, content type)``, or :class:`.FileStorage`.
+        Alternatively, pass raw bytes to set as :attr:`input_stream`, or an IO
+        object to read and set. ``json`` can be used to set JSON data instead.
+        ``content_length`` is set automatically.
     :param json: An object to be serialized and assigned to ``data``.
         Defaults the content type to ``"application/json"``.
         Serialized with the function assigned to :attr:`json_dumps`.
@@ -242,6 +221,13 @@ class EnvironBuilder:
     :param auth: An authorization object to use for the
         ``Authorization`` header value. A ``(username, password)`` tuple
         is a shortcut for ``Basic`` authorization.
+    :param input_stream: An IO object to pass through as the body of the
+        request, without reading, which simulates a streaming request. The
+        stream is not closed when calling :meth:`.close`, as it must remain open
+        to be read in the application.
+
+    .. versionchanged:: 3.2
+        Can be used as a ``with`` context manager to automatically close
 
     .. versionchanged:: 3.0
         The ``charset`` parameter was removed.
@@ -283,17 +269,14 @@ class EnvironBuilder:
     #: The default request class used by :meth:`get_request`.
     request_class = Request
 
-    import json
-
     #: The serialization function used when ``json`` is passed.
     json_dumps = staticmethod(json.dumps)
-    del json
 
-    _args: MultiDict[str, str] | None
-    _query_string: str | None
-    _input_stream: t.IO[bytes] | None
-    _form: MultiDict[str, str] | None
-    _files: FileMultiDict | None
+    _args: MultiDict[str, str] = MultiDict()
+    _query_string: str | None = None
+    _form: MultiDict[str, str] = MultiDict()
+    _files: FileMultiDict = FileMultiDict()
+    _input_stream: t.IO[bytes] | None = None
 
     def __init__(
         self,
@@ -309,7 +292,7 @@ class EnvironBuilder:
         multiprocess: bool = False,
         run_once: bool = False,
         headers: Headers | t.Iterable[tuple[str, str]] | None = None,
-        data: None | (t.IO[bytes] | str | bytes | t.Mapping[str, t.Any]) = None,
+        data: t.IO[bytes] | str | bytes | t.Mapping[str, t.Any] | None = None,
         environ_base: t.Mapping[str, t.Any] | None = None,
         environ_overrides: t.Mapping[str, t.Any] | None = None,
         mimetype: str | None = None,
@@ -353,7 +336,6 @@ class EnvironBuilder:
         self.environ_overrides = environ_overrides
         self.input_stream = input_stream
         self.content_length = content_length
-        self.closed = False
 
         if auth is not None:
             if isinstance(auth, tuple):
@@ -394,7 +376,7 @@ class EnvironBuilder:
             self.mimetype = mimetype
 
     @classmethod
-    def from_environ(cls, environ: WSGIEnvironment, **kwargs: t.Any) -> EnvironBuilder:
+    def from_environ(cls, environ: WSGIEnvironment, **kwargs: t.Any) -> te.Self:
         """Turn an environ dict back into a builder. Any extra kwargs
         override the args extracted from the environ.
 
@@ -527,96 +509,120 @@ class EnvironBuilder:
         else:
             self.headers["Content-Length"] = str(value)
 
-    def _get_form(self, name: str, storage: type[_TAnyMultiDict]) -> _TAnyMultiDict:
-        """Common behavior for getting the :attr:`form` and
-        :attr:`files` properties.
-
-        :param name: Name of the internal cached attribute.
-        :param storage: Storage class used for the data.
-        """
-        if self.input_stream is not None:
-            raise AttributeError("an input stream is defined")
-
-        rv = getattr(self, name)
-
-        if rv is None:
-            rv = storage()
-            setattr(self, name, rv)
-
-        return rv  # type: ignore
-
-    def _set_form(self, name: str, value: MultiDict[str, t.Any]) -> None:
-        """Common behavior for setting the :attr:`form` and
-        :attr:`files` properties.
-
-        :param name: Name of the internal cached attribute.
-        :param value: Value to assign to the attribute.
-        """
-        self._input_stream = None
-        setattr(self, name, value)
-
     @property
     def form(self) -> MultiDict[str, str]:
-        """A :class:`MultiDict` of form values."""
-        return self._get_form("_form", MultiDict)
+        """Form data text values. File values are stored in :attr:`files`.
+
+        If any values are set and no files are set, the request body will be
+        encoded and the content type set to ``application/x-www-form-urlencoded``.
+
+        Set when the constructor ``data`` parameter is a dict or multidict.
+
+        Cannot be accessed if :attr:`input_stream` is set. Setting this will
+        unset :attr:`input_stream`.
+        """
+        if self.input_stream is not None:
+            raise AttributeError("Not available when 'input_stream' is set.")
+
+        return self._form
 
     @form.setter
     def form(self, value: MultiDict[str, str]) -> None:
-        self._set_form("_form", value)
+        self._input_stream = None
+        self._form = value
 
     @property
     def files(self) -> FileMultiDict:
-        """A :class:`FileMultiDict` of uploaded files. Use
-        :meth:`~FileMultiDict.add_file` to add new files.
+        """Form data file values. Text values are stored in :attr:`form`.
+
+        If any values are set, the request body will be encoded and the content
+        type set to ``multipart/form-data``.
+
+        Set when the constructor ``data`` parameter is a dict or multidict.
+
+        The :meth:`.FileMultiDict.add_file` method provides a convenient way to
+        add more files without needing to construct :class:`.FileStorage`
+        objects.
+
+        The file streams will be read when encoding the data. All streams will
+        be closed when the builder's :meth:`.close` method is called.
+
+        Cannot be accessed if :attr:`input_stream` is set. Setting this will
+        unset :attr:`input_stream`.
+
+        Setting this will _not_ close files in the previous dict, call
+        :meth:`.FileMultiDict.close` first if that's needed.
         """
-        return self._get_form("_files", FileMultiDict)
+        if self.input_stream is not None:
+            raise AttributeError("Not available when 'input_stream' is set.")
+
+        return self._files
 
     @files.setter
     def files(self, value: FileMultiDict) -> None:
-        self._set_form("_files", value)
+        self._input_stream = None
+        self._files = value
 
     @property
     def input_stream(self) -> t.IO[bytes] | None:
-        """An optional input stream. This is mutually exclusive with
-        setting :attr:`form` and :attr:`files`, setting it will clear
-        those. Do not provide this if the method is not ``POST`` or
-        another method that has a body.
+        """A binary IO object to pass through as the body of the request,
+        without reading, which simulates a streaming request.
+
+        If this is set, :attr:`form` and :attr:`files` cannot be accessed.
+        If those are set, this will be ``None``. Setting this will close any
+        files and clear those.
+
+        The stream is not closed when calling the builder's :meth:`close`
+        method, as it must remain open to be read in the application.
+
+        .. versionchanged:: 3.2
+            Any values in :attr:`files` are closed first when setting this.
         """
         return self._input_stream
 
     @input_stream.setter
     def input_stream(self, value: t.IO[bytes] | None) -> None:
+        self._form.clear()
+        self._files.clear()
         self._input_stream = value
-        self._form = None
-        self._files = None
 
     @property
     def query_string(self) -> str:
-        """The query string.  If you set this to a string
-        :attr:`args` will no longer be available.
+        """The URL query string.
+
+        If this is set to a string, :attr:`args` cannot be accessed. If
+        :attr:`args` is set, this will be the encoded value of that. If neither
+        is set, this is the empty string.
         """
         if self._query_string is None:
             if self._args is not None:
                 return _urlencode(self._args)
+
             return ""
+
         return self._query_string
 
     @query_string.setter
     def query_string(self, value: str | None) -> None:
+        self._args.clear()
         self._query_string = value
-        self._args = None
 
     @property
     def args(self) -> MultiDict[str, str]:
-        """The URL arguments as :class:`MultiDict`."""
+        """The URL query string as a :class:`MultiDict`.
+
+        Set when the constructor ``query_string`` parameter is a dict or
+        multidict.
+
+        Setting this will unset :attr:`query_string`.
+        """
         if self._query_string is not None:
-            raise AttributeError("a query string is defined")
-        if self._args is None:
-            self._args = MultiDict()
+            raise AttributeError("Not available when 'query_string' is set.")
+
         return self._args
 
     @args.setter
-    def args(self, value: MultiDict[str, str] | None) -> None:
+    def args(self, value: MultiDict[str, str]) -> None:
         self._query_string = None
         self._args = value
 
@@ -641,23 +647,24 @@ class EnvironBuilder:
         return 80
 
     def __del__(self) -> None:
-        try:
-            self.close()
-        except Exception:
-            pass
+        self.close()
+
+    def __enter__(self) -> te.Self:
+        return self
+
+    def __exit__(
+        self,
+        exc_type: type[BaseException] | None,
+        exc_val: BaseException | None,
+        exc_tb: TracebackType | None,
+    ) -> None:
+        self.close()
 
     def close(self) -> None:
-        """Closes all files.  If you put real :class:`file` objects into the
-        :attr:`files` dict you can call this method to automatically close
-        them all in one go.
+        """Close all open files in :attr:`files`. :attr:`input_stream` is not
+        closed, as it is assumed to be managed externally.
         """
-        if self.closed:
-            return
-
-        if self._files is not None:
-            self.files.close()
-
-        self.closed = True
+        self._files.close()
 
     def get_environ(self) -> WSGIEnvironment:
         """Return the built environ.
@@ -1042,8 +1049,7 @@ class Client:
             if builder.input_stream is not None:
                 builder.input_stream.close()
 
-            builder.close()
-            builder.input_stream = None
+            builder.input_stream = None  # also closes and clears form and files
             builder.content_type = None
             builder.content_length = None
             builder.headers.pop("Content-Encoding", None)
@@ -1099,17 +1105,14 @@ class Client:
             if isinstance(arg, EnvironBuilder):
                 request = arg.get_request()
             elif isinstance(arg, dict):
-                request = EnvironBuilder.from_environ(arg).get_request()
+                with EnvironBuilder.from_environ(arg) as builder:
+                    request = builder.get_request()
             elif isinstance(arg, Request):
                 request = arg
 
         if request is None:
-            builder = EnvironBuilder(*args, **kwargs)
-
-            try:
+            with EnvironBuilder(*args, **kwargs) as builder:
                 request = builder.get_request()
-            finally:
-                builder.close()
 
         response_parts = self.run_wsgi_app(request.environ, buffered=buffered)
         response = self.response_wrapper(*response_parts, request=request)
@@ -1206,12 +1209,8 @@ def create_environ(*args: t.Any, **kwargs: t.Any) -> WSGIEnvironment:
        was added in 0.5.  The `headers`, `environ_base`, `environ_overrides`
        and `charset` parameters were added.
     """
-    builder = EnvironBuilder(*args, **kwargs)
-
-    try:
+    with EnvironBuilder(*args, **kwargs) as builder:
         return builder.get_environ()
-    finally:
-        builder.close()
 
 
 def run_wsgi_app(