+++ CHANGES.rst
@@ -26,6 +26,11 @@ Unreleased
     part of the HTTP spec. :pr:`3092`
 -   ``EnvironBuilder.close`` closes all open files in ``files`` rather than only
     the first for each key. :pr:`3092`
+-   ``EnvironBuilder`` can be used as a ``with`` context manager. :pr:`3101`
+-   ``EnvironBuilder.files.add_file`` will detect the filename when passing an
+    IO object. :pr:`3101`
+-   Added the ``EnvironBuilder.files.close`` method to close all files.
+    ``EnvironBuilder.files.clear`` will call ``close``. :pr:`3101`
 -   ``Map`` takes a ``subdomain_matching`` parameter to disable subdomain
     matching. In ``bind_to_environ``, the ``server_name`` parameter is not used
     if ``host_matching`` is enabled. If ``default_subdomain`` is set, it is used
@@ -34,6 +39,10 @@ Unreleased
     validated against ``request.trusted_hosts``. An invalid host will raise a
     400 error. :issue:`3007`
 -   Watchdog reloader is more efficient at ignoring events. :issue:`3090`
+-   If multipart parsing fails after some files have already been parsed, they
+    are closed to prevent a ``ResourceWarning``. :pr:`3101`
+-   ``SpooledTemporaryFile`` is always used for multipart file parsing.
+    :pr:`3101`
 
 
 Version 3.1.5